refactor: simplify subtitle store
