+++ frontend/src/store/client-settings/subtitle-settings.ts
@@ -1,15 +1,13 @@
-import { watchImmediate } from '@vueuse/core';
 import { watch } from 'vue';
 import { remote } from '@/plugins/remote';
 import { sealed } from '@/utils/validation';
 import { SyncedStore } from '@/store/super/synced-store';
-import { useFont } from '@/composables/use-font';
 
 /**
  * == INTERFACES AND TYPES ==
  */
 
-interface SubtitleApperance {
+export interface SubtitleSettingsState {
   fontFamily: string;
   fontSize: number;
   positionFromBottom: number;
@@ -17,49 +15,22 @@ interface SubtitleApperance {
   stroke: boolean;
 }
 
-export interface SubtitleSettingsState {
-  subtitleAppearance: SubtitleApperance;
-}
-
 @sealed
 class SubtitleSettingsStore extends SyncedStore<SubtitleSettingsState> {
-  private readonly _bodyFont = useFont();
-
-  public set subtitleAppearance(newVal: SubtitleSettingsState['subtitleAppearance']) {
-    this._state.subtitleAppearance = newVal;
-  }
-
-  public get subtitleAppearance(): SubtitleSettingsState['subtitleAppearance'] {
-    return this._state.subtitleAppearance;
-  }
-
-  private readonly _updateSubtitleFontFamily = (): void => {
-    this.subtitleAppearance.fontFamily
-        = this.subtitleAppearance.fontFamily === 'auto'
-        ? this._bodyFont.currentFont.value
-        : this.subtitleAppearance.fontFamily;
-  };
+  public state = this._state;
 
   public constructor() {
     super('subtitleSettings', {
-      subtitleAppearance: {
-        fontFamily: 'auto',
-        fontSize: 1.5,
-        positionFromBottom: 10,
-        backdrop: true,
-        stroke: false
-      }
+      fontFamily: 'auto',
+      fontSize: 1.5,
+      positionFromBottom: 10,
+      backdrop: true,
+      stroke: false
     }, 'localStorage');
 
     /**
      * == WATCHERS ==
      */
-
-    /**
-     * Font family changes
-     */
-    watchImmediate(this._bodyFont.currentFont, this._updateSubtitleFontFamily);
-
     watch(
       () => remote.auth.currentUser,
       () => {
@@ -71,4 +42,4 @@ class SubtitleSettingsStore extends SyncedStore<SubtitleSettingsState> {
   }
 }
 
-export default new SubtitleSettingsStore();
+export const subtitleSettings = new SubtitleSettingsStore();
