+++ frontend/src/store/super/common-store.ts
@@ -1,5 +1,5 @@
 import { useStorage, type RemovableRef } from '@vueuse/core';
-import { isRef, reactive } from 'vue';
+import { reactive, toValue } from 'vue';
 import { mergeExcludingUnknown } from '@/utils/data-manipulation';
 import { isNil } from '@/utils/validation';
 
@@ -11,7 +11,7 @@ export abstract class CommonStore<T extends object> {
   private readonly _internalState: T | RemovableRef<T>;
 
   protected get _state(): T {
-    return isRef(this._internalState) ? this._internalState.value : this._internalState;
+    return toValue(this._internalState);
   }
 
   protected readonly _reset = (): void => {
@@ -27,7 +27,7 @@ export abstract class CommonStore<T extends object> {
     if (persistence === 'localStorage') {
       storage = window.localStorage;
     } else if (persistence === 'sessionStorage') {
-      storage = sessionStorage;
+      storage = window.sessionStorage;
     }
 
     this._internalState = isNil(storage)
